KyberJS: move remaining tests to typescript
