+++ external/js/kyber/spec/constants.spec.ts
@@ -1,0 +1,7 @@
+import Constants from '../src/constants'
+
+describe('Kyber constants', () => {
+    it('should provide the 0 constant using big number', () => {
+        expect(Constants.zeroBN).toBeDefined();
+    });
+});
