+++ external/js/kyber/spec/group/edwards25519/curve.spec.ts
@@ -1,0 +1,27 @@
+import Curve from '../../../src/curve/edwards25519/curve';
+import Scalar from '../../../src/curve/edwards25519/scalar';
+import Point from '../../../src/curve/edwards25519/point';
+
+describe("edwards25519", () => {
+    const curve = new Curve();
+  
+    it("should return the name of the curve", () => {
+      expect(curve.string()).toBe("Ed25519", "Curve name is not correct");
+    });
+  
+    it("scalarLen should return the length of scalar", () => {
+      expect(curve.scalarLen()).toBe(32, "Scalar length not correct");
+    });
+  
+    it("pointLen should return the length of point", () => {
+      expect(curve.pointLen()).toBe(32, "Point length not correct");
+    });
+  
+    it("scalar should return a scalar", () => {
+      expect(curve.scalar()).toEqual(jasmine.any(Scalar));
+    });
+  
+    it("point should return a point", () => {
+      expect(curve.point()).toEqual(jasmine.any(Point));
+    });
+});
