+++ external/js/kyber/spec/group/edwards25519/sign.input
