+++ external/js/kyber/spec/index.spec.ts
@@ -1,0 +1,11 @@
+import { sign, curve } from '../src';
+
+describe('Kyber', () => {
+    it('should provide the curves', () => {
+        expect(curve).toBeDefined();
+    });
+
+    it('should provide the signatures', () => {
+        expect(sign).toBeDefined();
+    });
+});
